+++ test/ModelMixins/DiscretelyTimeVaryingMixinSpec.ts
@@ -43,4 +43,23 @@ describe("DiscretelyTimeVaryingMixin", () => {
     await wmsItem.loadChartItems();
     expect(wmsItem.chartItems[0].getColor()).toBe("#efefef");
   });
+
+  it("sets multiplier correctly from multiplierDefaultDeltaStep", async function() {
+    wmsItem = new WebMapServiceCatalogItem("mywms2", terria);
+    wmsItem.setTrait(
+      "definition",
+      "url",
+      "/test/WMS/period_datetimes_many_intervals.xml"
+    );
+
+    await wmsItem.loadMetadata();
+
+    expect(wmsItem.multiplier).toBeDefined();
+    expect(wmsItem.multiplierDefaultDeltaStep).toBeDefined();
+
+    // This dataset has a timestep every minute
+    expect(Math.round(wmsItem.multiplier!)).toBe(
+      60 / wmsItem.multiplierDefaultDeltaStep!
+    );
+  });
 });